Add CexAssetManager
