+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/cex/BalanceCexRepositoryWrapper.kt
@@ -1,18 +1,18 @@
 package io.horizontalsystems.bankwallet.modules.balance.cex
 
-import io.horizontalsystems.bankwallet.core.managers.MarketKitWrapper
-import io.horizontalsystems.bankwallet.core.providers.*
+import io.horizontalsystems.bankwallet.core.managers.CexAssetManager
+import io.horizontalsystems.bankwallet.core.providers.CexAsset
+import io.horizontalsystems.bankwallet.core.providers.ICexProvider
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
 
-class BalanceCexRepositoryWrapper(private val marketKit: MarketKitWrapper) {
+class BalanceCexRepositoryWrapper(private val cexAssetManager: CexAssetManager) {
     val itemsFlow = MutableStateFlow<List<CexAsset>?>(null)
 
     private var cexProvider: ICexProvider? = null
     private val coroutineScope = CoroutineScope(Dispatchers.Default)
     private var collectCexRepoItemsJob: Job? = null
-    private val coins = marketKit.allCoins().map { it.uid to it }.toMap()
 
     fun start() = Unit
 
@@ -22,15 +22,13 @@ class BalanceCexRepositoryWrapper(private val marketKit: MarketKitWrapper) {
 
     fun refresh() {
         collectCexRepoItemsJob?.cancel()
+        val provider = cexProvider ?: return
+
         collectCexRepoItemsJob = coroutineScope.launch {
             try {
-                val cexAssetRaws = cexProvider?.getAssets()
-                val cexAssets = cexAssetRaws?.map {
-                    buildCexAsset(it)
-                }
-
+                cexAssetManager.saveAll(provider.getAssets(), provider.account)
                 itemsFlow.update {
-                    cexAssets
+                    cexAssetManager.getAll(provider.account)
                 }
             } catch (t: Throwable) {
 
@@ -38,33 +36,14 @@ class BalanceCexRepositoryWrapper(private val marketKit: MarketKitWrapper) {
         }
     }
 
-    private fun buildCexAsset(cexAssetRaw: CexAssetRaw): CexAsset {
-        return CexAsset(
-            id = cexAssetRaw.id,
-            name = cexAssetRaw.name,
-            freeBalance = cexAssetRaw.freeBalance,
-            lockedBalance = cexAssetRaw.lockedBalance,
-            depositEnabled = cexAssetRaw.depositEnabled,
-            withdrawEnabled = cexAssetRaw.withdrawEnabled,
-            networks = cexAssetRaw.networks.map { buildCexNetwork(it) },
-            coin = coins[cexAssetRaw.coinUid],
-            decimals = cexAssetRaw.decimals
-        )
-    }
-
-    private fun buildCexNetwork(cexNetworkRaw: CexNetworkRaw): CexNetwork {
-        return CexNetwork(
-            network = cexNetworkRaw.network,
-            name = cexNetworkRaw.name,
-            isDefault = cexNetworkRaw.isDefault,
-            depositEnabled = cexNetworkRaw.depositEnabled,
-            withdrawEnabled = cexNetworkRaw.withdrawEnabled,
-            blockchain = null,
-        )
-    }
-
     fun setCexProvider(cexProvider: ICexProvider?) {
         this.cexProvider = cexProvider
+
+        itemsFlow.update {
+            cexProvider?.let {
+                cexAssetManager.getAll(it.account)
+            }
+        }
 
         refresh()
     }
