--- app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/cex/IBalanceCexRepository.kt
@@ -1,5 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.balance.cex
-
-interface IBalanceCexRepository {
-    suspend fun getItems(): List<BalanceCexItem>
-}
